requirements.txt
certifi==2022.5.18
chardet==4.0.0
decorator==5.1.1
idna==2.10
pika==1.3.1
Pillow==7.2.0
pkg-resources==0.0.0
py==1.11.0
requests==2.25.1
retry==0.9.2
urllib3==1.26.9
